requirements.txt
Flask==3.1.3
Flask-WTF==1.2.2
WTForms==3.2.1
psycopg2-binary==2.9.11
requests==2.33.1
feedparser==6.0.12
yfinance==1.2.2
beautifulsoup4==4.14.3
email-validator==2.3.0
azure-ai-inference==1.0.0b9
python-dotenv==1.0.1
gunicorn==23.0.0
